CMS-338: Update endpoints for publishing dates.
