+++ backend/index.js
@@ -12,6 +12,7 @@ import parkRoutes from "./routes/api/parks.js";
 import seasonRoutes from "./routes/api/seasons.js";
 import winterSeasonRoutes from "./routes/api/winter-seasons.js";
 import exportRoutes from "./routes/api/export.js";
+import PublishRoutes from "./routes/api/publish.js";
 
 if (!process.env.POSTGRES_SERVER || !process.env.ADMIN_PASSWORD) {
   throw new Error("Required environment variables are not set");
@@ -60,6 +61,7 @@ apiRouter.use("/parks", parkRoutes);
 apiRouter.use("/seasons", seasonRoutes);
 apiRouter.use("/export", exportRoutes);
 apiRouter.use("/winter-fees", winterSeasonRoutes);
+apiRouter.use("/publish", PublishRoutes);
 
 app.use("/api", checkJwt, apiRouter);
